infrastructure/functions/requirements.txt
functions_framework
google-cloud-storage
pandas